Move feature to separate folder; Add internalization
