+++ package.nls.json
@@ -129,6 +129,8 @@
   "param.enableCCache.title": "Enable CCache in build task",
   "param.enableSizeTaskAfterBuildTask.title": "Enable IDF Size task after build task",
   "param.sdkconfigFilePath.title": "sdkconfig file absolute path",
+  "param.hintsViewer.title": "Path to the hints file.",
+  "espIdf.searchError.title": "ESP-IDF: Search Error Hint",
   "esp.rainmaker.backend.sync.title": "Sync with ESP-Rainmaker Cloud Server",
   "esp.rainmaker.backend.connect.title": "Connect with ESP-Rainmaker Cloud Server",
   "esp.rainmaker.backend.logout.title": "Unlink Rainmaker Account",
