+++ src/extension.ts
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 "use strict";
-import * as os from 'os';
 import * as path from "path";
 import * as vscode from "vscode";
 import { srcOp, UpdateCmakeLists } from "./cmake/srcsWatcher";
@@ -98,7 +97,6 @@ import {
 import { generateConfigurationReport } from "./support";
 import { initializeReportObject } from "./support/initReportObj";
 import { writeTextReport } from "./support/writeReport";
-import { kill } from "process";
 import { getNewProjectArgs } from "./newProject/newProjectInit";
 import { NewProjectPanel } from "./newProject/newProjectPanel";
 import { buildCommand } from "./build/buildCmd";
@@ -134,8 +132,7 @@ import {
 } from "./project-conf";
 import { clearPreviousIdfSetups } from "./setup/existingIdfSetups";
 import { getEspRainmaker } from "./rainmaker/download/espRainmakerDownload";
-import * as yaml from 'js-yaml';
-import * as fs from 'fs';
+import { ErrorHintProvider } from "./espIdf/hints/index"
 
 // Global variables shared by commands
 let workspaceRoot: vscode.Uri;
@@ -3026,15 +3023,6 @@ export async function activate(context: vscode.ExtensionContext) {
           treeDataProvider.searchError(errorMsg, workspaceRoot);
       }
   });
-
-  vscode.commands.registerCommand('espIdf.promptErrorSearch', async () => {
-    const errorMsg = await vscode.window.showInputBox({
-      placeHolder: 'Enter the error message'
-    });
-    if (errorMsg) {
-      treeDataProvider.searchError(errorMsg, workspaceRoot);
-    }
-  });
 }
 
 function validateInputForRainmakerDeviceParam(
@@ -3554,144 +3542,4 @@ class IdfDebugConfigurationProvider
     }
     return config;
   }
-}
-
-class ReHintPair {
-    re: string;
-    hint: string;
-    match_to_output: boolean;
-
-    constructor(re: string, hint: string, match_to_output: boolean = false) {
-        this.re = re;
-        this.hint = hint;
-        this.match_to_output = match_to_output;
-    }
-}
-
-class ErrorHint {
-  public type: 'error' | 'hint';
-  public children: ErrorHint[] = [];
-
-  constructor(public label: string, type: 'error' | 'hint') {
-      this.type = type;
-  }
-
-  addChild(child: ErrorHint) {
-      this.children.push(child);
-  }
-}
-
-class ErrorHintProvider implements vscode.TreeDataProvider<ErrorHint> {
-	constructor(private context: vscode.ExtensionContext) { }
-	private _onDidChangeTreeData: vscode.EventEmitter<ErrorHint | undefined | null | void> = new vscode.EventEmitter<ErrorHint | undefined | null | void>();
-	readonly onDidChangeTreeData: vscode.Event<ErrorHint | undefined | null | void> = this._onDidChangeTreeData.event;
-
-	private data: ErrorHint[] = [];
-
-  searchError(errorMsg: string, workspace) {
-    console.log(`Searching for: ${errorMsg}`);
-    const espIdfPath = idfConf.readParameter("idf.espIdfPath", workspace) as string;
-    const hintsPath = getHintsYmlPath(espIdfPath);
-    const fileContents = fs.readFileSync(hintsPath, 'utf-8');
-    const hintsData = yaml.load(fileContents);
-
-    const reHintsPairArray: ReHintPair[] = this.loadHints(hintsData);
-
-    this.data = [];
-    for (const hintPair of reHintsPairArray) {
-    const match = new RegExp(hintPair.re).exec(errorMsg);
-    console.log(`RegExp matched: ${JSON.stringify(match)}`);
-    if (match) {
-        let finalHint = hintPair.hint;
-
-        if (hintPair.match_to_output && hintPair.hint.includes("{}")) {
-            // Replace {} with the first capturing group from the regex match
-            finalHint = hintPair.hint.replace("{}", match[0]);
-        }
-
-        const error = new ErrorHint(hintPair.re, 'error');
-        const hint = new ErrorHint(finalHint, 'hint');
-        error.addChild(hint); 
-        this.data.push(error);
-        console.log(`Found error: ${hintPair.re}`);
-        console.log(`Found hint: ${finalHint}`);
-    }
-}
-
-    if (!this.data.length) {
-      console.log('No hints found');
-      this.data.push(new ErrorHint(`No hints found for ${errorMsg}`, 'error'));
-    }
-
-    this._onDidChangeTreeData.fire();
-  }
-
-  private loadHints(hintsArray: any): ReHintPair[] {
-    let reHintsPairArray: ReHintPair[] = [];
-
-    for (const entry of hintsArray) {
-        if (entry.variables && entry.variables.length) {
-            for (const variableSet of entry.variables) {
-                const reVariables = variableSet.re_variables;
-                const hintVariables = variableSet.hint_variables;
-
-                let re = this.formatEntry(reVariables, entry.re);
-                let hint = this.formatEntry(hintVariables, entry.hint);
-
-                reHintsPairArray.push(new ReHintPair(re, hint, entry.match_to_output));
-            }
-        } else {
-            let re = String(entry.re);
-            let hint = String(entry.hint);
-
-            if (!entry.match_to_output) {
-              re = this.formatEntry([], re);
-              hint = this.formatEntry([], hint);
-            }
-
-            reHintsPairArray.push(new ReHintPair(re, hint, entry.match_to_output));
-        }
-    }
-
-    return reHintsPairArray;
-} 
-
-  private formatEntry(vars: string[], entry: string): string {
-      let i = 0;
-      while (entry.includes("{}")) {
-          entry = entry.replace("{}", "{" + i++ + "}");
-      }
-      const result = entry.replace(/\{(\d+)\}/g, (_, idx) => vars[Number(idx)] || "");
-      return result;
-  }
-
-  getTreeItem(element: ErrorHint): vscode.TreeItem {
-    let treeItem = new vscode.TreeItem(element.label);
-
-    if (element.type === 'error') {
-        if (element.label.startsWith("No hints found")) {
-          treeItem.label = `⚠️ ${element.label}`
-        } else {
-            treeItem.label = `🔍 ${element.label}`;
-            treeItem.collapsibleState = vscode.TreeItemCollapsibleState.Expanded; // Ensure errors are expanded by default
-        }
-    } else if (element.type === 'hint') {
-        treeItem.label = `💡 ${element.label}`;
-    }
-
-    return treeItem;
-  }
-
-  getChildren(element?: ErrorHint): Thenable<ErrorHint[]> {
-    if (element) {
-        return Promise.resolve(element.children); // Return children if there's a parent element
-    } else {
-        return Promise.resolve(this.data);
-    }
-  }
-}
-
-function getHintsYmlPath(espIdfPath: string): string {
-  const separator = os.platform() === 'win32' ? '\\' : '/';
-  return `${espIdfPath}${separator}tools${separator}idf_py_actions${separator}hints.yml`;
 }
